+++ src/components/data-browser/index.ts
@@ -1,0 +1,5 @@
+export { DatabaseSidebar } from './database-sidebar';
+export { MainContent } from './main-content';
+export { GridView } from './grid-view';
+export { ListView } from './list-view';
+export { TableView } from './table-view';
